Get species occurrences from the ALA or GBIF

A simple wrapper to occurrences from ALA4R. Performs a search for a species (e.g. 'Eucalyptus globulus'), sets the download_reason to 'testing' (see ?ala_reasons), and returns only the species, the longitude, and the latitude of all observations.

get_occurrences_ala(species, ala_args = NULL)

get_occurrences_gbif(species, gbif_args = NULL)

Arguments

species
The full scientific species name, quoted.

ala_args
Optionally, a list of arguments passed to occurrences (ALA4R package).

gbif_args
Optionally, a list of arguments passed to occ_search (rgbif package).

Details

Note that when passing additional arguments to occ_search, the following arguments are already set and can not be overriden : hasGeospatialIssue (FALSE), limit (50000), fields (species, lat and long), and of course scientificName. When passing additional arguments to occurrences, only download_reason (7) is already set.

See also

get_occurrences_gbif

Examples

# NOT RUN {
o <- get_occurrences_ala("Eucalyptus botryoides")

# Should return a message like:
# INFO [2017-05-12 09:52:10] ALA returned 3274 records for Eucalyptus botryoides in 2.6 sec.
# }