Download Worldclim rasters

Downloads and unzips 10min resolution Worldlclim rasters, stores them for reuse.

get_worldclim_rasters(wc_vars = "tavg")

Arguments

wc_vars
The WorldClim variables to be downloaded, e.g. 'tavg', 'prec' (see get_climate_vars).

Details

Zip files are downloaded from http://biogeo.ucdavis.edu/data/worldclim/v2.0/tif/base. This function is normally not invoked by the user, since climate_presence as well as get_climate_vars download the rasters as needed. The path to download the rasters to is read from options()$worldclimpath.

Examples

# Return worldclim tmean and prec rasters, as a list with those components

# NOT RUN {
wc <- get_worldclim_rasters(wc_vars=c("tavg","tmin"))
# }